wip: stuff and important bugfix
In the main loop when checking if a song is currently played in order to
update the LCD, the connection established by `mpd_connection_new` has
not be freed when no song is currently playing (since the call to
`mpd_connection_free` was located inside the if-clause only executed
when a "current song" is present). This caused mpd to exceed the max
connections limit after short and thus not accepting any further
connections, making the MPDaemon non-responding und effectively useless.